fix issue with map resizing when layer panel is hidden in ResilienceMapCard
