+++ src/components/Map/MapCard.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 import PropTypes from "prop-types";
 import { useSelector } from "react-redux";
 import { useMapEvents } from "react-leaflet";
@@ -9,8 +9,20 @@ import BasemapLayer from "./BasemapLayer.jsx";
 export default function MapCard({ children, map, setMap, mapEventHandlers }) {
   const selectedCenterSelector = (state) => state.mapProperties.center;
   const selectedZoomSelector = (state) => state.mapProperties.zoom;
+  const listVisibleSelector = (state) => state.mapLayerList.visible;
+  const areaVisibleSelector = (state) => state.mapProperties.areaVisible;
   const center = useSelector(selectedCenterSelector, () => true);
   const zoom = useSelector(selectedZoomSelector, () => true);
+  const layerListVisible = useSelector(listVisibleSelector);
+  const areaVisible = useSelector(areaVisibleSelector);
+
+  useEffect(() => {
+    let timer;
+    if (map) {
+      timer = setTimeout(() => map.invalidateSize(), 10);
+    }
+    return () => clearTimeout(timer);
+  }, [map, areaVisible, layerListVisible]);
 
   // This component exists solely for the useMapEvents hook
   const MapEventsComponent = () => {
